+++ Sources/Logger.swift
@@ -12,14 +12,28 @@
 //
 
 import Foundation
+import OSLog
 
-/// The default logger implementation.
-final class Logger: Logging {
+/// Describes a type that logs messages for the library.
+public protocol Logging: Sendable {
 
+    /// Logs the provided message.
+    ///
+    /// - Parameter message: The message to log.
+    func log(_ message: @escaping @autoclosure () -> String)
+}
+
+/// A default ``Logging`` implementation to log debug messages.
+///
+/// You can set this logger for the ``CollectionViewDriver.logger``.
+public final class RCKLogger: Logging {
     /// The shared logger instance.
-    static let shared = Logger()
+    public static let shared = RCKLogger()
 
-    func log(_ message: @autoclosure () -> String) {
-        print(message())
+    private let _logger = Logger(subsystem: "com.jessesquires.ReactiveCollectionsKit", category: "")
+
+    /// :nodoc:
+    public func log(_ message: @escaping @autoclosure () -> String) {
+        self._logger.debug("\(message())")
     }
 }